repository_miner/requirements.txt
flask
GitPython
requests

